refactoring IdleTimeout component
