+++ packages/design-system/src/components/IdleTimeout/IdleTimeout.test.tsx
@@ -16,42 +16,77 @@ describe('Idle Timeout', () => {
     return render(<IdleTimeout {...defaultProps} {...overrideProps} />);
   };
 
-  describe('timeout countdown', async () => {
+  beforeEach(() => {
+    jest.useFakeTimers();
+  });
+
+  afterEach(() => {
+    jest.useRealTimers();
+  });
+
+  it('should error if timeToWarning is greater than timeToTimeout', () => {
+    const error = jest.spyOn(console, 'error').mockImplementation(() => {});
+    renderIdleTimeout({ timeToWarning: 7 });
+    expect(error).toHaveBeenCalledWith(
+      'Error in TimeoutManager component. `timeToWarning` is greater or equal to `timeToTimeout`'
+    );
+    error.mockReset();
+  });
+
+  describe('timeout countdown', () => {
     // user explicitly says "keep session"
     xit('should reset countdown if user opts for that', () => {});
+
+    it('warning element should be visible at set warning time', () => {
+      const { queryByTestId } = renderIdleTimeout();
+      const warningEl = queryByTestId('warning-element-mock');
+      expect(warningEl).toBeNull();
+      jest.advanceTimersByTime(timeTilWarningShown);
+      expect(warningEl).toBeDefined();
+    });
+
+    xit('should reset countdown if mouse moves', () => {});
+
+    xit('should reset countdown if key is pressed', () => {});
+
+    it('should call onTimeout when countdown ends', () => {
+      renderIdleTimeout();
+      jest.advanceTimersByTime(defaultProps.timeToTimeout * 60000);
+      expect(defaultProps.onTimeout).toHaveBeenCalled();
+    });
   });
 
   describe('action buttons', () => {
     it('should call onTimeout if onSessionForcedEnd is not provided', () => {
       const { getByText } = renderIdleTimeout({ showSessionEndButton: true });
-      //   jest.advanceTimersByTime(timeTilWarningShown);
+      jest.advanceTimersByTime(timeTilWarningShown);
       const endSessionBtn = getByText('Logout');
       fireEvent.click(endSessionBtn);
       expect(onTimeout).toHaveBeenCalled();
     });
 
     it('should close warning modal when user opts to continue session', () => {
-      const { getByText, getByRole } = renderIdleTimeout();
-      // jest.advanceTimersByTime(timeTilWarningShown);
+      const { getByText, queryByRole } = renderIdleTimeout();
+      jest.advanceTimersByTime(timeTilWarningShown);
       const keepSessionBtn = getByText('Continue session');
       fireEvent.click(keepSessionBtn);
-      const dialogEl = getByRole('alertdialog');
-      expect(dialogEl).toBeUndefined();
+      const dialogEl = queryByRole('alertdialog');
+      expect(dialogEl).toBeNull();
     });
   });
 
   it('should replace token in message', () => {
     const message = 'Your session will end in <timeToTimeout>.';
     const { getByRole } = renderIdleTimeout({ message });
-    // jest.advanceTimersByTime(timeTilWarningShown);
+    jest.advanceTimersByTime(timeTilWarningShown);
     const dialogBodyText = getByRole('main');
-    expect(dialogBodyText.textContent).toEqual('Your session will end in 5 minutes.');
+    expect(dialogBodyText.textContent).toEqual('Your session will end in 3 minutes.');
   });
 
   it('should adjust message for singular minute vs multiple', () => {
     const message = 'Your session will end in <timeToTimeout>.';
     const { getByRole } = renderIdleTimeout({ message, timeToWarning: 1 });
-    // jest.advanceTimersByTime(timeTilWarningShown);
+    jest.advanceTimersByTime((defaultProps.timeToTimeout - 1) * 60000);
     const dialogBodyText = getByRole('main');
     expect(dialogBodyText.textContent).toEqual('Your session will end in 1 minute.');
   });
@@ -59,10 +94,10 @@ describe('Idle Timeout', () => {
   it('should replace multiple token instances in message', () => {
     const message = 'Your session will end in <timeToTimeout>. <timeToTimeout> until session ends';
     const { getByRole } = renderIdleTimeout({ message });
-    // jest.advanceTimersByTime(timeTilWarningShown);
+    jest.advanceTimersByTime(timeTilWarningShown);
     const dialogBodyText = getByRole('main');
     expect(dialogBodyText.textContent).toEqual(
-      'Your session will end in 5 minutes. 5 minutes until session ends'
+      'Your session will end in 3 minutes. 3 minutes until session ends'
     );
   });
