+++ packages/design-system/src/components/IdleTimeout/index.ts
@@ -1,3 +1,2 @@
 export * from './IdleTimeout';
 export * from './IdleTimeoutDialog';
-export * from './TimeoutManager';
